remove BlogPageContainer and 'connect' pagination with browserHistory
